requirements.txt
mock
pyserial>=3.2.1
pytest
